Move all source files into a subdirectory to try to keep things somewhat separated.
We may want to alter the directory structure further (e.g. separate out spectre_* -- the actual runnable demonstrations -- from instr and cache_sidechannel and so on), but this is a decent first step.

The name "demos" was picked arbitrarily -- I don't generally see "demonstrations" used as a word, and "examples" seems to imply to me that they are examples of use of an API, instead of examples of Spectre and so on.

Change-Id: Idd62f6ee9455eb5f80b3c454cbcf7201f49142e0