Fixing CMake on PowerPC and Windows.
Spectre v1 BTB does not work on PowerPC with -O1. It requires -O2.
Windows examples are slow and unreliable without Release mode and full optimizations.
Meltdown works on ppc64le. I don't know whether it works on ppc64 and unless we have tested it, we should not publish it.
Meltdown kernel module should be built iff meltdown is built.