Fixing CMake on PowerPC and Windows.
Windows examples are slow and unreliable without Release mode and full optimizations. Even on the other systems there is no reason to default to less optimized versions.
Meltdown works on ppc64le. I don't know whether it works on ppc64 and unless we have tested it, we should not publish it.
Meltdown kernel module should be built iff meltdown is built.