Always use "${}" in ifs()
There are a few ways to refer to variables in CMake `if` commands:

1. `if(VARIABLE ...)`
2. `if(${VARIABLE} ...)`
3. `if("${VARIABLE}" ...)`

These forms all work wherever the command accepts `<variable|string>`, but
it turns out (1) and (2) can have surprising behavior.

Form (1) evaluates to the contents of `VARIABLE`, if it is defined, otherwise
it evaluates to _the string_ `"VARIABLE"`.

Form (2) evaluates to the contents of `VARIABLE`, unless that value itself
happens to be the name of another variable, in which case _that_ variable's
value is used.

Only form (3) always evaluates to either the contents of `VARIABLE` or empty.

ref: https://cmake.org/cmake/help/v3.16/command/if.html#variable-expansion