Allow a small number of false positives in `timing_array_test`
We are seeing very occasional false positives in CI. My best explanation is that there is still a prefetcher operating that does not respect page boundaries. Of course, it's also very possible there's a bug I'm missing.

For now, the error rate is still low enough for our channel to be useful. Update `timing_array_test` to make sure it stays that way.