Widen timing difference
Increase optimization to `-O3`. This probably reduces memory contention from other loads and stores that are now optimized away.

Move the call to `std::fill` to its own noinline function. Without this change, at `-O3`, the code takes *longer* to run and shows *no* timing difference across the values 0, 1, and 2. My guess is that the `"memory"` clobber in `CompilerOpaqueUse` was having effects on other variables that needed to be spilled to memory.

```
Fill with 0: 45us
Fill with 1: 51us
Fill with 2: 51us
```