Fix builds+CI when kernel headers are unavailable
Install kernel headers opportunistically on Linux hosts, and allow builds on Linux to continue even if headers aren't installed. In that case, the kernel modules will be skipped.

This fixes CI on ppc64le, where the build runs an Ubuntu image in an LXD container and the host is running a kernel with no corresponding headers package in the Ubuntu repositories.