More clearly define randomness in timing_array.h
This is a Linear Congruential Generator. See e.g. https://en.cppreference.com/w/cpp/numeric/random/linear_congruential_engine and compare the formula ;)  i.e. we're basically reimplementing `std::linear_congruential_engine<113, 100, 256>`

This is an important observation, because it implies further restrictions on the number 113 -- it shouldn't just be odd, but in fact *prime* (or at least co-prime with the modulus), and 112 needs to satisfy several tests (in particular, divisible by 2 and 4, but not 8). This also informs us how we can extend this to arbitrary sizes rather than 256.

The most obvious next step, if we want to reuse this for non-256-sized arrays, is to make it a PCG over the next-biggest sized power of 2 (and step the PCG until an in-range value is found). Alternatively, any other PRNG that supports O(1) seek and can have its period set arbitrarily small would suffice.

I'm not sure what these are called in computer science -- in math, it's just called a permutation. As in, I want a permutation over the integers 0..n -- a bijection from 0..n to 0..n. But if I google for permutation functions, I get algorithms for shuffling vectors instead of int->int permutations. :/