Replace shuffled vector with an LCG-based shuffled access pattern.
This should reduce/remove measurement artifacts from the shuffle vector.
Since the shuffle vector is, itself, the same size as the buffer being
tested for cache, it can lower the measured value by a factor of 2. In
fact, it can reduce it by more than 2, because the buffer is of char and
the shuffle vector is of int32 or even int64 -- meaning it could cut it
by a whole order of magnitude!